pyproject.toml
[project]
name = "flashresearch"
version = "0.2.0"
description = "A deep research agent built with LangGraph with RACE/FACT evaluation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["langgraph", "research", "agent", "llm", "tavily"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-tavily>=0.1.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
flashresearch = "run:main"
flashresearch-benchmark = "benchmark:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flashresearch"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
